backend/pyproject.toml
[project]
name = "catchlog-backend"
version = "0.1.0"
description = "CatchLog fishing compliance API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "python-dotenv>=1.2.1",
    "pydantic-ai>=1.63.0",
    "boto3>=1.42.59",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
]
